// server/src/core/env-key.enum.ts
export enum EnvKey {
	APP_HOST = 'APP_HOST',
	APP_PORT = 'APP_PORT',
	APP_FRONTEND_URL = 'APP_FRONTEND_URL',
	
	DB_HOST = 'DB_HOST',
	DB_PORT = 'DB_PORT',
	DB_USER = 'DB_USER',
	DB_PASSWORD = 'DB_PASSWORD',
	DB_DEFAULT = 'DB_DEFAULT',
	DB_LOGGING = 'DB_LOGGING',
	
	SMTP = 'SMTP',
	EMAIL = 'EMAIL',
	PASSWORD = 'PASSWORD',
	
	JWT_SECRET = 'JWT_SECRET',
}